refactor(app): split top-level into multiple submodules
