+++ crates/app/src/lib.rs
@@ -5,12 +5,78 @@
 #![forbid(unsafe_code)]
 #![feature(str_split_as_str)]
 
-mod app;
+mod builder;
+mod handle;
 mod repos;
 mod tags;
 mod trees;
 
-pub use app::*;
+pub use builder::*;
+pub(crate) use handle::*;
 pub use repos::*;
 pub use tags::*;
 pub use trees::*;
+
+use std::fmt::Display;
+use std::sync::Arc;
+
+use drawbridge_store::{Get, Memory};
+use drawbridge_type::{RepositoryName, TagName, TreePath};
+
+use axum::http::StatusCode;
+use tokio::sync::RwLock;
+
+pub type RepoStore = RwLock<Memory<RepositoryName>>;
+pub type TagStore = RwLock<Memory<(RepositoryName, TagName)>>;
+pub type TreeStore = RwLock<Memory<(RepositoryName, TagName, TreePath)>>;
+
+pub async fn assert_repo(
+    repos: Arc<RepoStore>,
+    repo: RepositoryName,
+) -> Result<(), (StatusCode, &'static str)> {
+    #[inline]
+    fn err_map(e: impl Display) -> (StatusCode, &'static str) {
+        eprintln!("failed to check repository existence: {}", e);
+        (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            "Failed to check repository existence",
+        )
+    }
+    if !repos
+        .read()
+        .await
+        .contains(repo.clone())
+        .await
+        .map_err(err_map)?
+    {
+        Err((StatusCode::NOT_FOUND, "Repository does not exist"))
+    } else {
+        Ok(())
+    }
+}
+
+pub async fn assert_tag(
+    tags: Arc<TagStore>,
+    repo: RepositoryName,
+    tag: TagName,
+) -> Result<(), (StatusCode, &'static str)> {
+    #[inline]
+    fn err_map(e: impl Display) -> (StatusCode, &'static str) {
+        eprintln!("failed to check tag existence: {}", e);
+        (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            "Failed to check tag existence",
+        )
+    }
+    if !tags
+        .read()
+        .await
+        .contains((repo.clone(), tag.clone()))
+        .await
+        .map_err(err_map)?
+    {
+        Err((StatusCode::NOT_FOUND, "Tag does not exist"))
+    } else {
+        Ok(())
+    }
+}
